refactor(test) : 테스트 코드 및 패키지명 수정
