+++ src/test/java/site/coach_coach/coach_coach_server/routine/controller/RoutineControllerTest.java
@@ -27,7 +27,7 @@ import site.coach_coach.coach_coach_server.auth.userdetails.CustomUserDetails;
 import site.coach_coach.coach_coach_server.routine.dto.RoutineForListDto;
 import site.coach_coach.coach_coach_server.routine.dto.RoutineListRequest;
 import site.coach_coach.coach_coach_server.routine.dto.UserInfoForRoutineList;
-import site.coach_coach.coach_coach_server.routine.services.RoutineService;
+import site.coach_coach.coach_coach_server.routine.service.RoutineService;
 
 @WebMvcTest(RoutineController.class)
 public class RoutineControllerTest {
